apps/backend/pyproject.toml
[tool.poetry]
name = "elect-backend"
version = "0.1.0"
description = "FastAPI backend for Elect 2027 - Kenyan voter information platform"
authors = ["Elect Team <team@elect2027.ke>"]
license = "MIT"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}
alembic = "^1.13.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
python-cors = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.25.0"
black = "^23.12.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
